requirements.txt
airflow==2.7.0  # แนะนำเวอร์ชันที่รองรับ
pandas
requests
python-dotenv

# certifi==2024.7.4
# charset-normalizer==3.3.2
# idna==3.7
# numpy==2.0.1
# pandas==2.2.2
# python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
# pytz==2024.1
# requests==2.32.3
# six==1.16.0
# tzdata==2024.1
# urllib3==2.2.2
